+++ app/store/index.js
@@ -1,5 +1,12 @@
+const mongodb = require("mongodb");
+
 const users = require("./modules/users");
+const photoSets = require("./modules/photosets");
+const photos = require("./modules/photos");
 
 module.exports = {
-    users
+    ObjectID: mongodb.ObjectID,
+    users,
+    photoSets,
+    photos
 };